D3D12 update

+++ Engine/Source/Rendering/D3D12/CommandBuffer.h
@@ -27,14 +27,14 @@ namespace Nova::D3D12
         void SetScissor(int32_t x, int32_t y, int32_t width, int32_t height) override;
         void Draw(uint32_t vertexCount, uint32_t instanceCount, uint32_t firstVertex, uint32_t firstInstance) override;
         void DrawIndexed(uint32_t indexCount, uint32_t instanceCount, uint32_t firstIndex, int32_t vertexOffset, uint32_t firstInstance) override;
-        void DrawIndirect(const Nova::Buffer& buffer, size_t offset, uint32_t drawCount) override;
+        void DrawIndirect(const Nova::Buffer& buffer, uint64_t offset, uint32_t drawCount) override;
         void DrawIndexedIndirect(const Nova::Buffer& buffer, uint64_t offset, uint32_t drawCount) override;
         void BeginRenderPass(const RenderPassBeginInfo& beginInfo) override;
         void EndRenderPass() override;
         void PushConstants(const Nova::Shader& shader, ShaderStageFlags stageFlags, size_t offset, size_t size,const void* values) override;
         void UpdateBuffer(const Nova::Buffer& buffer, size_t offset, size_t size, const void* data) override;
         void Dispatch(uint32_t groupCountX, uint32_t groupCountY, uint32_t groupCountZ) override;
-        void DispatchIndirect(const Nova::Buffer& buffer, size_t offset) override;
+        void DispatchIndirect(const Nova::Buffer& buffer, uint64_t offset) override;
         void BufferCopy(const Nova::Buffer& src, const Nova::Buffer& dest, size_t srcOffset, size_t destOffset,size_t size) override;
         void Blit(const Nova::Texture& src, const Nova::BlitRegion& srcRegion, const Nova::Texture& dest, const Nova::BlitRegion& destRegion, Filter filter) override;
         void Blit(const Nova::Texture& src, const Nova::Texture& dest, Filter filter) override;
