+++ Engine/Source/Rendering/Vulkan/CommandBuffer.h
@@ -31,7 +31,7 @@ namespace Nova::Vulkan
         void SetViewport(float x, float y, float width, float height, float minDepth, float maxDepth) override;
         void SetScissor(int32_t x, int32_t y, int32_t width, int32_t height) override;
         void Draw(uint32_t vertexCount, uint32_t instanceCount, uint32_t firstVertex, uint32_t firstInstance) override;
-        void DrawIndirect(const Nova::Buffer& buffer, size_t offset, uint32_t drawCount) override;
+        void DrawIndirect(const Nova::Buffer& buffer, uint64_t offset, uint32_t drawCount) override;
         void DrawIndexed(uint32_t indexCount, uint32_t instanceCount, uint32_t firstIndex, int32_t vertexOffset, uint32_t firstInstance) override;
         void DrawIndexedIndirect(const Nova::Buffer& buffer, uint64_t offset, uint32_t drawCount) override;
         void BeginRenderPass(const Nova::RenderPassBeginInfo& beginInfo) override;
@@ -43,7 +43,7 @@ namespace Nova::Vulkan
         void MemoryBarrier(const Nova::MemoryBarrier& memoryBarrier) override;
         
         void Dispatch(uint32_t groupCountX, uint32_t groupCountY, uint32_t groupCountZ) override;
-        void DispatchIndirect(const Nova::Buffer& buffer, size_t offset) override;
+        void DispatchIndirect(const Nova::Buffer& buffer, uint64_t offset) override;
 
         void BufferCopy(const Nova::Buffer& src, const Nova::Buffer& dest, size_t srcOffset, size_t destOffset, size_t size) override;
         void CopyBufferToTexture(const Nova::Buffer& src, const Nova::Texture& dest, size_t srcOffset, size_t srcSize, uint32_t arrayIndex, uint32_t mipLevel) override;
